Working Business JPA. Failed to implement Lazy Loading of Business Photo.

+++ src/main/java/dev/group21/rescmeal/repository/BusinessRepository.java
@@ -3,6 +3,8 @@ package dev.group21.rescmeal.repository;
 import dev.group21.rescmeal.model.Business;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessRepository extends JpaRepository<Business, Integer> {
+public interface BusinessRepository extends JpaRepository<Business, Long> {
     // Queries
+//    @Query("SELECT b FROM Business b LEFT JOIN FETCH business_photo WHERE b.id_business = business_photo.id_business")
+//    Optional<Business> findByIdWithPhoto(@Param("id_business") Integer id_business);
 }
